requirements.txt
streamlit
google-generativeai
gspread
google-auth
google-api-python-client
pandas